Create TabControllerProps namespace
