+++ generatedTypes/components/tabController/TabPage.d.ts
@@ -1,5 +1,6 @@
 import React, { PureComponent } from 'react';
 import Reanimated from 'react-native-reanimated';
+import _ from 'lodash';
 export interface TabPageProps {
     /**
      * The index of the the TabPage
@@ -65,6 +66,6 @@ export default class TabPage extends PureComponent<TabPageProps> {
         width?: undefined;
     } | undefined)[];
     lazyLoad: () => void;
-    renderCodeBlock: () => Reanimated.Node<number>;
+    renderCodeBlock: (() => JSX.Element) & _.MemoizedFunction;
     render(): JSX.Element;
 }
