+++ generatedTypes/index.d.ts
@@ -42,7 +42,7 @@ export {default as Overlay, OverlayTypes} from './components/overlay';
 export {default as RadioButton, RadioButtonPropTypes, RadioButtonProps} from './components/radioButton/RadioButton';
 export {default as RadioGroup, RadioGroupPropTypes, RadioGroupProps} from './components/radioButton/RadioGroup';
 export {default as Switch, SwitchProps} from './components/switch';
-export {default as TabController, TabControllerProps, TabBarItemProps} from './components/tabController';
+export {TabController, TabControllerProps} from './components/tabController';
 export {default as TabBar, TabBarProps} from './components/TabBar';
 export {default as Fader, FaderProps, FaderPosition} from './components/fader';
 export {default as ExpandableSection, ExpandableSectionProps} from './components/expandableSection';
