+++ src/components/tabController/index.ts
@@ -1,0 +1,7 @@
+import {default as TabController, TabControllerProps as _TabControllerProps, TabBarItemProps as _TabBarItemProps} from './TabController';
+declare namespace Props {
+    type TabControllerProps = _TabControllerProps;
+    type TabBarItemProps = _TabBarItemProps;
+}
+
+export {TabController, Props as TabControllerProps};
