+++ src/index.js
@@ -148,7 +148,7 @@ export default {
     return require('./components/stepper').default;
   },
   get TabController() {
-    return require('./components/tabController').default;
+    return require('./components/tabController').TabController;
   },
   get TabBar() {
     return require('./components/tabBar').default;
